refactor: inject Redis client externally
